requirements.txt
streamlit==1.43.2
openai>=1.68.2
chromadb==0.5.23
pydantic==2.10.6
python-dotenv==1.0.1
pypdf==5.4.0
markdown==3.7
